requirements.txt
python-decouple
pyrogram
tgcrypto
heroku3
psycopg2
gitpython
asyncio